grpc_tuto/pyproject.toml
[tool.poetry]
name = "grpc_tuto"
version = "0.1.0"
description = ""
authors = ["sebhmg <sebastienh@mirageoscience.com>"]

[tool.poetry.dependencies]
python = "^3.7"
grpcio-tools = "^1.23"
numpy = "^1.17"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
